Move parsing modules into a folder
